fix: relocating dapppnode hook
