+++ features/dashboard/bond/last-rewards.tsx
@@ -17,7 +17,6 @@ import {
   getPrevRewardsFrame,
   useLastOperatorRewards,
   useLastRewardsSlot,
-  useLastRewrdsTx,
   useNodeOperatorInfo,
 } from 'shared/hooks';
 import { formatDate, formatPercent } from 'utils';
@@ -29,6 +28,7 @@ import {
   RowHeader,
   RowTitle,
 } from './styles';
+import { useLastRewrdsTx } from 'dappnode/hooks/useLastRewardsFrame-api'; // DAPPNODE
 
 export const LastRewards: FC = () => {
   const { data: lastRewards, initialLoading: isLoading } =
